Added interrupts

+++ src/lib/inc/def.h
@@ -13,8 +13,8 @@ typedef unsigned char u8;
 typedef signed short s16;
 typedef unsigned short u16;
 
-typedef signed int s32;
-typedef unsigned int u32;
+typedef signed long s32;
+typedef unsigned long u32;
 
 typedef signed long long s64;
 typedef unsigned long long u64;
